minor fxp changes, print out -3dB close up

+++ dfilt_fxp/dfilt_fxp.c
@@ -17,9 +17,9 @@ dfilt_single_pole_u16(uint16_t input, uint16_t prev_output, uint16_t alpha)
   uint16_t output;                  /* current time step output */
   uint32_t beta = ((1<<15)-alpha);  /* coefficient of previous output */
   /* calculate the previous output term and scale to output scaling */
-  uint32_t y1 = (beta*prev_output) >> 15;
+  uint16_t y1 = (beta*(uint32_t)prev_output) >> 15;
   /* calculate the input term and scale to output scaling */
-  uint32_t u1 = (alpha*input) >> 15;
+  uint16_t u1 = ((uint32_t)alpha*(uint32_t)input) >> 15;
 	/* calculate current filter output */
   output = y1 + u1;
 	return output;
